+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/extension/ResourceAdapter.java
@@ -18,7 +18,9 @@ package io.fabric8.kubernetes.client.extension;
 import io.fabric8.kubernetes.api.builder.Visitor;
 import io.fabric8.kubernetes.api.model.DeletionPropagation;
 import io.fabric8.kubernetes.api.model.ListOptions;
+import io.fabric8.kubernetes.api.model.PartialObjectMetadata;
 import io.fabric8.kubernetes.api.model.StatusDetails;
+import io.fabric8.kubernetes.api.model.Table;
 import io.fabric8.kubernetes.api.model.autoscaling.v1.Scale;
 import io.fabric8.kubernetes.client.GracePeriodConfigurable;
 import io.fabric8.kubernetes.client.PropagationPolicyConfigurable;
@@ -386,4 +388,13 @@ public class ResourceAdapter<T> implements Resource<T> {
     return resource.subresource(subresource);
   }
 
+  @Override
+  public PartialObjectMetadata getAsPartialObjectMetadata() {
+    return resource.getAsPartialObjectMetadata();
+  }
+
+  @Override
+  public Table getAsTable() {
+    return resource.getAsTable();
+  }
 }